pyproject.toml
[project]
name = "codebase_analysis"
requires-python = ">=3.8"
version = "0.0.1"
dependencies = [
    "pandas==2.2.3",
    "numpy==2.2.3",
    "openai==1.64.0",
    "streamlit==1.43.0",
    "PyYAML==6.0.2",
    "psycopg2-binary==2.9.10",
    "GitPython==3.1.44",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.dynamic]
readme = {file = ["README.md"], content-type = "text/markdown"}

[tool.ruff]
exclude = ["*.ipynb"]

[tool.ruff.lint]
ignore = ["E501", "F401"]